+++ R/survalis-package.R
@@ -84,5 +84,7 @@ utils::globalVariables(c(
   "splits","ale","integrated_ale","type","change_label","hjust",
   # new in 1.1.0: dca_survmat/roc_survmat/plot_rmst/plot_shap_beeswarm
   "threshold","net_benefit","strategy","specificity","sensitivity",
-  "tau","estimate","lo","hi","scaled_value"
+  "tau","estimate","lo","hi","scaled_value",
+  # new: compute_survcluster/plot_survcluster
+  "survival","subject","cluster"
 ))